issue #176 reefactor: move FormModal directory and remove unused BaseModal isBackdropClose prop and wrong usage
