+++ src/components/Common/Modal/BaseModal.jsx
@@ -8,16 +8,13 @@ import { CloseIcon } from "@/constants/iconConstants.js";
 import { closeModal } from "@/features/ui/ui-slice.js";
 import useScrollLock from "@/hooks/useScrollLock.jsx";
 
-import { BackdropDiv, IconButton, ModalDiv } from "./BaseModal.style.js";
+import { BackdropDiv, IconButton, ModalDiv } from "./BaseModal.styles.js";
 
-const Backdrop = ({ isBackdropClose }) => {
+const Backdrop = () => {
 	const dispatch = useDispatch();
 
 	return (
-		<BackdropDiv
-			className="backdrop"
-			onClick={() => isBackdropClose && dispatch(closeModal())}
-		/>
+		<BackdropDiv className="backdrop" onClick={() => dispatch(closeModal())} />
 	);
 };
 
@@ -38,15 +35,12 @@ const Modal = ({ children }) => {
 	);
 };
 
-const BaseModal = ({ children, isBackdropClose }) => {
+const BaseModal = ({ children }) => {
 	useScrollLock();
 
 	return (
 		<>
-			{ReactDOM.createPortal(
-				<Backdrop isBackdropClose={isBackdropClose} />,
-				document.getElementById("backdrop"),
-			)}
+			{ReactDOM.createPortal(<Backdrop />, document.getElementById("backdrop"))}
 			{ReactDOM.createPortal(
 				<Modal>{children}</Modal>,
 				document.getElementById("modal"),
