+++ src/components/Setting/GroupModals/GroupModals.jsx
@@ -23,10 +23,10 @@ import {
 	TitleHeader,
 } from "./GroupModals.style";
 
-const modalStyle = {
-	padding: "20px",
-	backgroundColor: "white",
-};
+// const modalStyle = {
+// 	padding: "20px",
+// 	backgroundColor: "white",
+// };
 
 export const GroupDeleteModal = ({ groupInfo, isLoading }) => {
 	const { groupId, name } = groupInfo;
@@ -44,7 +44,8 @@ export const GroupDeleteModal = ({ groupInfo, isLoading }) => {
 	};
 
 	return (
-		<BaseModal style={modalStyle}>
+		<BaseModal>
+			{/* <BaseModal style={modalStyle}> */}
 			<ContainerDiv>
 				<TitleHeader>
 					<strong>{`${name}을(를) 정말 삭제하실 건가요?`}</strong>
@@ -112,7 +113,8 @@ export const GroupDelegateModal = ({ groupInfo, isGroupLoading }) => {
 	}, []);
 
 	return (
-		<BaseModal style={modalStyle}>
+		<BaseModal>
+			{/* <BaseModal style={modalStyle}> */}
 			<ContainerDiv>
 				<TitleHeader>
 					<strong>{`${name}을(를) 위임받을 그룹원을 선택해주세요`}.</strong>
@@ -168,7 +170,8 @@ export const GroupLeaveModal = ({ groupInfo, isLoading }) => {
 	};
 
 	return (
-		<BaseModal style={modalStyle}>
+		<BaseModal>
+			{/* <BaseModal style={modalStyle}> */}
 			<ContainerDiv>
 				<TitleHeader>
 					<strong>{`${name}을(를) 정말 나가실 건가요?`}</strong>
