+++ src/components/Setting/WithdrawalTab/WithdrawalTab.jsx
@@ -61,13 +61,14 @@ const WithdrawalTab = () => {
 				</WithdrawalButton>
 			</ButtonWrapDiv>
 			{openedModal === "WITHDRAW" && (
-				<BaseModal
+				<BaseModal>
+					{/* <BaseModal
 					style={{
 						backgroundColor: "white",
 						borderRadius: "10px",
 					}}
 					hasClose={false}
-				>
+				> */}
 					<ModalContentDiv>
 						<p>
 							Selody를 정말
